chore(errors): moved errors into separate package
